debian/local/plymouth.hook: Apply module exclusions for MODULES=dep as well
The Plymouth initramfs-tools hook avoids adding certain DRM modules to
the initramfs, although it's not clear exactly why. Right now this is
only done if MODULES=most in initramfs-tools config, but they should be
excluded when MODULES=dep as well.

There are more modules we can want to exclude in both cases, for example
drivers like Panfrost that only do 3D rendering acceleration which are
not necessary to output an image to the display.

Extract the modules-to-exclude to a variable we can use with either
initramfs-tools configuration. While adding modules for the DRM devices
on the running system, check if they are one of these modules and skip
those that are.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>