d/p/ubuntu-spinner-fsck.patch: listen to /run/systemd/fsck.progress
When systemd-fsckd was removed from upstream systemd, a small
helper/compatibility feature was left behind. When systemd-fsck is
running, it will connect to /run/systemd/fsck.progress, if it exists,
and point fsck to that fd for progress updates. The systemd-fsckd patch
in Ubuntu currently utilizes that, and then sends updates to plymouth.
This means that we currently have a situation like this:

 systemd-fsck@dev1.service -
 ...                        |-> systemd-fsckd.service -> plymouthd
 systemd-fsck@devN.service -

In other words, systemd-fsckd.service is a middleware. But, plymouthd
could simply listen on /run/systemd/fsck.progress itself. That is what
this patch introduces.

The ubuntu-spinner-fsck patch already adds special logic to plymouth for
fsck progress. Add additional logic to plymouthd which listens on this socket,
and then does similar formatting to what systemd-fsckd would do. Instead
of sending updates via plymouth's command protocol, just call the update
handler directly with messages like "fsckd:<num_devs>:<min_progress>".

This means the existing format of messages received from systemd-fsckd
won't be supported anymore, but that's OK because systemd-fsckd will be
removed from Ubuntu.