d/rules, d/install: Restore plymouth config files
Ubuntu uses alternatives to configure the plymouth theme rather than the Theme=
option in the config file. Because of that the config files were removed from
the install.

However, the config files can accept other options. Plymouth will misbehave
if a file of configuration defaults is not provided. Restore the configuration
files, and just remove the Theme= option to keep using the alternatives logic.

This is shipped as a patch rather than a debian/rules scriptlet to ensure it
will be noticed in case of changes, in the form of a merge conflict.